services/api_fastapi/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["routers", "auth", "storage", "models", "config.py", "main.py"]

[project]
name = "api-fastapi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "google-cloud-firestore>=2.14,<3.0",
    "firebase-admin>=6.5,<7.0",
    "python-dotenv>=1.0,<2.0",
    "structlog>=24.1,<25.0",
    "scoring-engine",
]

[tool.uv.sources]
scoring-engine = { path = "../scoring_engine" }

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
